+++ tket/src/Clifford/UnitaryTableau.cpp
@@ -499,6 +499,11 @@ UnitaryRevTableau::UnitaryRevTableau(const qubit_vector_t& qbs) : tab_(qbs) {}
 
 UnitaryRevTableau::UnitaryRevTableau() : tab_() {}
 
+UnitaryRevTableau::UnitaryRevTableau(
+    const MatrixXb& xx, const MatrixXb& xz, const VectorXb& xph,
+    const MatrixXb& zx, const MatrixXb& zz, const VectorXb& zph)
+    : tab_(xx, xz, xph, zx, zz, zph) {}
+
 SpPauliStabiliser UnitaryRevTableau::get_xrow(const Qubit& qb) const {
   return tab_.get_xrow(qb);
 }
@@ -594,16 +599,16 @@ std::ostream& operator<<(std::ostream& os, const UnitaryRevTableau& tab) {
   for (unsigned i = 0; i < nqs; ++i) {
     Qubit qi = tab.tab_.qubits_.right.at(i);
     os << tab.tab_.tab_.xmat_.row(i) << "   " << tab.tab_.tab_.zmat_.row(i)
-       << "   " << tab.tab_.tab_.phase_(i) << "\t->\t"
-       << "X@" << qi.repr() << std::endl;
+       << "   " << tab.tab_.tab_.phase_(i) << "\t->\t" << "X@" << qi.repr()
+       << std::endl;
   }
   os << "--" << std::endl;
   for (unsigned i = 0; i < nqs; ++i) {
     Qubit qi = tab.tab_.qubits_.right.at(i);
     os << tab.tab_.tab_.xmat_.row(i + nqs) << "   "
        << tab.tab_.tab_.zmat_.row(i + nqs) << "   "
-       << tab.tab_.tab_.phase_(i + nqs) << "\t->\t"
-       << "Z@" << qi.repr() << std::endl;
+       << tab.tab_.tab_.phase_(i + nqs) << "\t->\t" << "Z@" << qi.repr()
+       << std::endl;
   }
   return os;
 }